+++ src/extension.js
@@ -10,7 +10,6 @@ function activate(context) {
   const channel = vscode.window.createOutputChannel(`${name}`);
   channel.appendLine(`${name}: activate`);
   let config = vscode.workspace.getConfiguration("boundaryMove");
-  let jumpZoomOutLevel = config.get("jumpZoomOutLevel", 1);
   const bm = new boundary.BoundaryManager(channel, config);
 
   vscodeUtil.registerCommand(context, "BM.moveLeft", moveLeft);
@@ -20,13 +19,14 @@ function activate(context) {
   vscodeUtil.registerCommand(context, "BM.info", info);
   vscodeUtil.registerCommand(context, "BM.jump", jump);
   vscodeUtil.registerCommand(context, "BM.jumpLine", jumpLine);
+  vscodeUtil.registerCommand(context, "BM.selectJump", selectJump);
+  vscodeUtil.registerCommand(context, "BM.selectJumpLine", selectJumpLine);
 
   vscode.workspace.onDidChangeConfiguration(
     (event) => {
       if (event.affectsConfiguration("boundaryMove") === true) {
         config = vscode.workspace.getConfiguration("boundaryMove");
         bm.config(config);
-        jumpZoomOutLevel = config.get("jumpZoomOutLevel", 1);
       }
     },
     null,
@@ -40,6 +40,7 @@ function activate(context) {
     null,
     context.subscriptions
   );
+
   vscode.workspace.onDidCloseTextDocument(
     (document) => {
       bm.delete(document);
@@ -47,6 +48,7 @@ function activate(context) {
     null,
     context.subscriptions
   );
+
   vscode.workspace.onDidChangeTextDocument(
     (event) => {
       bm.change(event);
@@ -91,164 +93,41 @@ function activate(context) {
     bm.info();
   }
 
-  async function jump() {
+  function jump() {
     const editor = vscode.window.activeTextEditor;
     if (vscodeUtil.isEmpty(editor) === true) {
       return;
     }
-
-    if (jumpZoomOutLevel > 0) {
-      await vscodeUtil.fontZoomOut(jumpZoomOutLevel);
-    }
-    const { documentIndex, start, end } = bm.getVisibleRange(editor);
-    if (documentIndex === -1 || end - start <= 0) {
-      if (jumpZoomOutLevel > 0) {
-        await vscodeUtil.fontZoomIn(jumpZoomOutLevel);
-      }
-      return;
-    }
-    const decorationRanges = bm.getDecorationRanges(documentIndex, start, end);
-    const decorationTypes = setDecorations(editor, decorationRanges);
-    const result = await showBoundaryInputRange(decorationRanges);
-    decorationTypes.dispose();
-    if (jumpZoomOutLevel > 0) {
-      await vscodeUtil.fontZoomIn(jumpZoomOutLevel);
-    }
-    if (result === null) {
-      return;
-    }
-
-    bm.jump(editor, documentIndex, result.lineIndex, result.count);
-    return;
+    bm.jump(editor);
   }
 
-  async function jumpLine() {
+  function selectJump() {
     const editor = vscode.window.activeTextEditor;
     if (vscodeUtil.isEmpty(editor) === true) {
       return;
     }
+    bm.jump(editor, true);
+  }
 
-    const { documentIndex, lineCount } = bm.getLineCount(editor);
-    if (documentIndex === -1 || lineCount <= 0) {
+  function jumpLine() {
+    const editor = vscode.window.activeTextEditor;
+    if (vscodeUtil.isEmpty(editor) === true) {
       return;
     }
-    if (jumpZoomOutLevel > 0) {
-      await vscodeUtil.fontZoomOut(jumpZoomOutLevel);
-    }
-    const lineIndex = await showLineInput(lineCount, editor.selection.active.line);
-    if (lineIndex === -1) {
-      if (jumpZoomOutLevel > 0) {
-        await vscodeUtil.fontZoomIn(jumpZoomOutLevel);
-      }
+    bm.jumpLine(editor);
+  }
+
+  function selectJumpLine() {
+    const editor = vscode.window.activeTextEditor;
+    if (vscodeUtil.isEmpty(editor) === true) {
       return;
     }
-
-    const lineDecorationRanges = bm.getLineDecorationRanges(documentIndex, lineIndex);
-    const decorationTypes = setDecorations(editor, [lineDecorationRanges]);
-    const count = await showBoundaryInput(lineDecorationRanges);
-    decorationTypes.dispose();
-    if (jumpZoomOutLevel > 0) {
-      await vscodeUtil.fontZoomIn(jumpZoomOutLevel);
-    }
-    if (count === -1) {
-      return;
-    }
-
-    bm.jump(editor, documentIndex, lineIndex, count);
-  }
-
-  /**
-   * @param {number} lineCount
-   */
-  async function showLineInput(lineCount, currntLine) {
-    const rangeString = `range: 1 -- ${lineCount}`;
-    const line = (currntLine + 1).toString(10);
-    const result = await vscode.window.showInputBox({
-      placeHolder: rangeString,
-      prompt: `Input line index for jump. ${rangeString}`,
-      value: line,
-      valueSelection: [0, line.length],
-    });
-    if (vscodeUtil.isEmpty(result) === true) {
-      return -1;
-    }
-    let lineIndex = parseInt(result, 10) - 1;
-    if (lineIndex < 0) {
-      lineIndex = 0;
-    } else if (lineIndex >= lineCount) {
-      lineIndex = lineCount - 1;
-    }
-    return lineIndex;
-  }
-
-  /**
-   * @param {{range: import("vscode").Range, textContent: string, index: number}[]} lineDecorationRanges
-   */
-  async function showBoundaryInput(lineDecorationRanges) {
-    const rangeString = `range: a -- ${bm.converter.convertToString(lineDecorationRanges.length - 1)}`;
-    const result = await vscode.window.showInputBox({
-      placeHolder: rangeString,
-      prompt: `Input boundary index for jump. ${rangeString}`,
-      value: "a",
-    });
-    if (vscodeUtil.isEmpty(result) === true) {
-      return -1;
-    }
-    let i = bm.converter.convertToNumber(result);
-    let count = 0;
-    if (i < 0) {
-      count = lineDecorationRanges[0].index;
-    } else if (i >= lineDecorationRanges.length) {
-      count = lineDecorationRanges[lineDecorationRanges.length - 1].index;
-    } else {
-      count = lineDecorationRanges[i].index;
-    }
-    return count;
-  }
-
-  async function showBoundaryInputRange(decorationRanges) {
-    const rangeString = `range: aa -- ${decorationRanges[decorationRanges.length - 1][decorationRanges[decorationRanges.length - 1].length - 1].textContent}`;
-    const input = await vscode.window.showInputBox({
-      placeHolder: rangeString,
-      prompt: `Input boundary index for jump. ${rangeString}`,
-    });
-    if (vscodeUtil.isEmpty(input) === true) {
-      return null;
-    }
-    return bm.search(decorationRanges, input);
-  }
-
-  /**
-   * @param {import("vscode").TextEditor} editor
-   * @param {{range: import("vscode").Range, textContent: string, index: number}[][]} decoratonRanges
-   */
-  function setDecorations(editor, decoratonRanges) {
-    const decorationType = vscode.window.createTextEditorDecorationType({});
-    const options = [];
-    for (const lineDecorationRanges of decoratonRanges) {
-      for (const option of lineDecorationRanges) {
-        options.push({
-          range: option.range,
-          renderOptions: {
-            before: {
-              contentText: option.textContent,
-              margin: bm.MarkerMargin,
-              fontStyle: "normal",
-              fontWeight: "normal",
-              color: { id: "boundaryMove.markerColor" },
-              backgroundColor: { id: "boundaryMove.markerBackgroundColor" },
-            },
-          },
-        });
-      }
-    }
-    editor.setDecorations(decorationType, options);
-    return decorationType;
+    bm.jumpLine(editor, true);
   }
 }
 exports.activate = activate;
 
-function deactivate() { }
+function deactivate() {}
 
 module.exports = {
   activate,