+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## [0.3.0] 2020/08/31
+
+### [Added]
+
+- Add selectJump command.
+- Add selectJumpLine command.
+- Add alwaysCenter.
+- Add jumpToCenter.
+
+## [0.2.4] 2020/08/23
+
+### [Fixed]
+
+- Fix concat.
+
 ## [0.2.3] 2020/08/23
 
 ### [Fixed]
@@ -48,7 +63,7 @@
 
 ### [Changed]
 
-- Chnage JumpMarker style
+- Change JumpMarker style
 - Add target file scheme
 
 ### [Fixed]